# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault IPTV API — activation and device management
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ActivationResponseStatus = typeof ActivationResponseStatus[keyof typeof ActivationResponseStatus];


export const ActivationResponseStatus = {
  active: 'active',
  pending: 'pending',
} as const;

export interface ActivationResponse {
  status: ActivationResponseStatus;
  /**
     * xtream or m3u
     * @nullable
     */
  type?: string | null;
  /** @nullable */
  host?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  m3u_url?: string | null;
}

export interface Device {
  id: number;
  mac_address: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  type?: string | null;
  /** @nullable */
  host?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  m3u_url?: string | null;
  created_at: string;
}

export type DeviceInputType = typeof DeviceInputType[keyof typeof DeviceInputType];


export const DeviceInputType = {
  xtream: 'xtream',
  m3u: 'm3u',
} as const;

export interface DeviceInput {
  mac_address: string;
  name?: string;
  type: DeviceInputType;
  host?: string;
  username?: string;
  password?: string;
  m3u_url?: string;
}

export type DeviceUpdateType = typeof DeviceUpdateType[keyof typeof DeviceUpdateType];


export const DeviceUpdateType = {
  xtream: 'xtream',
  m3u: 'm3u',
} as const;

export interface DeviceUpdate {
  name?: string;
  type?: DeviceUpdateType;
  host?: string;
  username?: string;
  password?: string;
  m3u_url?: string;
}

export interface ErrorResponse {
  error: string;
}

export type CheckActivationParams = {
/**
 * Device MAC address (e.g. AA:BB:CC:DD:EE:FF)
 */
mac: string;
};