# lib/api-zod/src/generated/types/activationResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault IPTV API — activation and device management
 * OpenAPI spec version: 0.1.0
 */
import type { ActivationResponseStatus } from './activationResponseStatus';

export interface ActivationResponse {
  status: ActivationResponseStatus;
  /**
     * xtream or m3u
     * @nullable
     */
  type?: string | null;
  /** @nullable */
  host?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  password?: string | null;
  /** @nullable */
  m3u_url?: string | null;
}