# lib/api-zod/src/generated/types/deviceInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault IPTV API — activation and device management
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceInputType } from './deviceInputType';

export interface DeviceInput {
  mac_address: string;
  name?: string;
  type: DeviceInputType;
  host?: string;
  username?: string;
  password?: string;
  m3u_url?: string;
}