# lib/api-zod/src/generated/types/deviceUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault IPTV API — activation and device management
 * OpenAPI spec version: 0.1.0
 */
import type { DeviceUpdateType } from './deviceUpdateType';

export interface DeviceUpdate {
  name?: string;
  type?: DeviceUpdateType;
  host?: string;
  username?: string;
  password?: string;
  m3u_url?: string;
}