# lib/api-zod/src/generated/types/errorResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault IPTV API — activation and device management
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorResponse {
  error: string;
}